# Add concurrency safety