--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,8 @@
+module github.com/getlantern/preconn
+
+go 1.16
+
+require (
+	github.com/getlantern/nettest v1.0.0
+	github.com/stretchr/testify v1.7.0
+)