In working on https://github.com/getlantern/tlsmasq/issues/20, I was running into data races rooted in this library.  It's a pretty easy and low-impact fix.

I benchmarked this change as follows:
```
func BenchmarkRead(b *testing.B) {
	const bufSize = 1024

	c1, c2 := net.Pipe()
	c1 = Wrap(c1, []byte("hello from c2"))

	data := make([]byte, bufSize*b.N)
	if _, err := rand.Reader.Read(data); err != nil {
		b.Fatalf("failed to prepare data: %v", err)
	}

	buf := make([]byte, bufSize)
	go io.Copy(c2, bytes.NewReader(data))

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		c1.Read(buf)
	}
}
```

On this branch, this produces:
```
goos: darwin
goarch: amd64
pkg: github.com/getlantern/preconn
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkRead-12    	  964540	      1118 ns/op
PASS
ok  	github.com/getlantern/preconn	3.574s
```

On master, this produces:
```
goos: darwin
goarch: amd64
pkg: github.com/getlantern/preconn
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkRead-12    	 1000000	      1088 ns/op
PASS
ok  	github.com/getlantern/preconn	3.284s
```

So, these changes constitute a slowdown of ~30 ns per operation, a difference of about 3%.