@oxtoacart 2021-06-16
Code LGTM, though I'm curious what scenario caused you to encounter this? In general, whether sockets or file descriptors, it's unusual/hazardous to have multiple readers from the same stream (though I could see it working with a PacketConn).