@hwh33 2021-06-16
Thanks @oxtoacart!

> I'm curious what scenario caused you to encounter this?

I'm adding `nettest.TestConn` to tlsmasq (https://github.com/getlantern/tlsmasq/issues/20).  [One of the subtests](https://cs.opensource.google/go/x/net/+/04defd46:nettest/conntest.go;l=347-393) specifically checks to ensure that all methods on the connection can be called concurrently without creating a race.  Other subtests, such as [this one](https://cs.opensource.google/go/x/net/+/04defd46:nettest/conntest.go;l=145-171), trigger data race errors in `preconn.Read` for incidental reasons.  I agree that you generally do not want unsynchronized concurrent readers on the same connection.

This particular fix probably does not resolve any bugs we'd run into in practice.  However, I think the benefit of `nettest.TestConn` (and the race detector) outweighs the downside of making little fixes for sort-of-bugs like this.