@hwh33 2021-06-17
Great idea on using `nettest` in this repo.  I added your test [here](https://github.com/getlantern/preconn/pull/4/commits/c7170b66e1f8c5cb25a14f7191810c2e559038ae).  This involved also adding a `go.mod` file to this repo.